refactor: move detectBuldedFunctions to vercel service
